requirements.txt
streamlit==1.39.0
PyMuPDF==1.24.9
requests==2.32.3
beautifulsoup4==4.12.3
openai==1.14.3
google-generativeai==0.8.3
sentence-transformers==3.2.1
faiss-cpu==1.9.0
langchain==0.1.16
langchain-openai==0.1.3
langchain-community==0.0.38
python-dotenv==1.0.1